+++ include/common.h
@@ -21,7 +21,9 @@ typedef void UNK_PTR;
 #include "sys/gfx/animation.h"
 #include "sys/gfx/model.h"
 #include "sys/audio/audio.h"
+#include "sys/exception.h"
 #include "sys/crash.h"
+#include "sys/asset_thread.h"
 #include "game/gamebits.h"
 #include "game/actor/objdata.h"
 #include "game/actor/hitbox.h"
