+++ include/sys/exception.h
@@ -1,5 +1,5 @@
-#ifndef _EXCEPTION_H_
-#define _EXCEPTION_H_
+#ifndef _SYS_EXCEPTION_H_
+#define _SYS_EXCEPTION_H_
 
 #include "common.h"
 
@@ -52,4 +52,9 @@ void some_crash_print(OSThread**, int, int);
  */
 void check_video_mode_crash_and_clear_framebuffer();
 
-#endif
+void clear_framebuffer_current();
+void pi_manager_entry(void *arg);
+void crash_controller_getter();
+void check_video_mode_crash_and_clear_framebuffer();
+
+#endif //_SYS_EXCEPTION_H_
