+++ src/input.c
@@ -1,254 +1,4 @@
 #include "common.h"
-#include "input.h"
-
-// NOTE: This size is NOT CONFIRMED YET
-#define CONTROLLER_THREAD_STACKSIZE 1152
-
-// The length of gContInterruptBuffer
-#define CONT_INTERRUPT_BUFFER_LENGTH 2
-// The length of gContThreadMesgQueueBuffer
-#define CONT_THREAD_MESG_QUEUE_BUFFER_LENGTH 8
-// The length of gContThreadInputsAppliedQueueBuffer
-#define CONT_THREAD_INPUTS_APPLIED_QUEUE_BUFFER_LENGTH 1
-
-// The maximum number of buffered controller snapshots.
-#define MAX_BUFFERED_CONT_SNAPSHOTS 4
-
-/**
- * Represents a single snapshot of each controller.
- * 
- * In addition to raw OSContPad snapshots, includes bitfields for
- * buttons being pressed and released.
- */
-typedef struct _ControllersSnapshot {
-    /**
-     * A raw controller pad snapshot for each controller.
-     */
-    OSContPad pads[MAXCONTROLLERS];
-
-    /**
-     * For each controller, a button bitfield where set button bits indicate that
-     * the button was just pressed down.
-     */
-    u16 buttonPresses[MAXCONTROLLERS];
-
-    /**
-     * For each controller, a button bitfield where set button bits indicate that
-     * the button was just released.
-     */
-    u16 buttonReleases[MAXCONTROLLERS];
-} ControllersSnapshot;
-
-/**
- * If TRUE, no controllers are inserted.
- * 
- * @see init_controller_data
- */
-extern s32 gNoControllers;
-
-extern u16 D_8008C8A4;
-
-/**
- * For each controller, a mask applied to their button state.
- * 
- * Ex. If the mask is 0x0000 then all buttons will be reported as 0 
- * regardless of whether they are actually pressed.
- * 
- * At the end of each input frame (see controller_thread_entry) this 
- * is reset to 0xFFFF for each controller.
- */
-extern u16 gButtonMask[MAXCONTROLLERS];
-
-extern u8 gIgnoreJoystick;
-
-/**
- * An index of gContSnapshots specifying which snapshot buffer contains previous controller snapshots.
- * 
- * The index of current controller snapshots can be found by doing: gPrevContSnapshotsI ^ 1.
- */
-extern u8 gPrevContSnapshotsI;
-
-/**
- * Whether the controller thread should apply controller inputs to globals like gContPads
- * after collecting controller interrupts.
- * 
- * @see gContThreadMesgQueue and controller_thread_entry for more info.
- */
-extern u8 gApplyContInputs;
-
-/**
- * For the current and previous controller snapshots, the number of snapshots that are buffered.
- * 
- * @details Which index is which is determined by gPrevContSnapshotsI.
- */
-extern u8 gNumBufContSnapshots[2];
-
-/**
- * Pointers to 2 arrays, where one is an array of current controller snapshots 
- * and the other is an array of previous controller snapshots.
- * 
- * @details Which array is currently which is determined by gPrevContSnapshotsI.
- * Each array is a buffer of up to 4 snapshots. This is done to avoid missing controller
- * interrupt data as it comes in on the queue, by being able to collect more than one at
- * a time when processing input.
- * 
- * @see controller_thread_entry
- */
-extern ControllersSnapshot *gContSnapshots[2];
-
-/**
- * A buffer of controller snapshots, gContSnapshots[0] is a pointer to this.
- */
-extern ControllersSnapshot gContSnapshotsBuffer0[MAX_BUFFERED_CONT_SNAPSHOTS];
-/**
- * A buffer of controller snapshots, gContSnapshots[1] is a pointer to this.
- */
-extern ControllersSnapshot gContSnapshotsBuffer1[MAX_BUFFERED_CONT_SNAPSHOTS];
-
-/**
- * The message queue associated with SI controller interrupts.
- * 
- * This is the same message queue that is passed to osContInit.
- */
-extern OSMesgQueue gContInterruptQueue;
-/**
- * A message queue for signalling when the controller thread should process controller
- * interrupts and potentially to apply them to the controller globals that the game 
- * reads from for normal gameplay.
- * 
- * - For every non-1 value sent to this queue, the controller thread will set a flag saying
- * that next time the queue gets a value of 1 sent, it will process controller interrupts
- * AND apply them to globals like gContPads.
- * - For every value of 1 sent to this queue, the controller thread will process controller
- * interrupts and if a non-1 value was previously sent also apply them to globals like gContPads.
- * 
- * While messages are not sent to this queue, the controller thread will block.
- * 
- * Usually, 0xA, 0x1 is repeatedly sent each frame or so (not sure exactly when, but its related
- * to game ticks) to process controller input.
- * 
- * See controller_thread_entry for details.
- */
-extern OSMesgQueue gContThreadMesgQueue;
-/**
- * When the controller thread applies controller inputs to input globals used by normal 
- * gameplay functions, a message is sent to this queue to signal this.
- */
-extern OSMesgQueue gContThreadInputsAppliedQueue;
-
-extern OSMesg gContInterruptBuffer[CONT_INTERRUPT_BUFFER_LENGTH];
-extern OSMesg gContThreadMesgQueueBuffer[CONT_THREAD_MESG_QUEUE_BUFFER_LENGTH];
-extern OSMesg gContThreadInputsAppliedQueueBuffer[CONT_THREAD_INPUTS_APPLIED_QUEUE_BUFFER_LENGTH];
-
-/**
- * Message sent to the gContInterruptQueue when an SI interrupt occurs.
- */
-extern OSMesg gContInterruptMessage;
-
-/**
- * The status and type of each controller (SI device).
- */
-extern OSContStatus gContStatuses[MAXCONTROLLERS];
-
-/**
- * Current inputs of each controller.
- * 
- * @details This data is created by combining multiple buffered controller snapshots 
- * from gContSnapshots.
- */
-extern OSContPad gContPads[MAXCONTROLLERS];
-
-/**
- * For each controller, a bitfield of buttons that were just pressed.
- */
-extern u16 gButtonPresses[MAXCONTROLLERS];
-/**
- * For each controller, a bitfield of buttons that were just released.
- */
-extern u16 gButtonReleases[MAXCONTROLLERS];
-
-/**
- * Maps virtual controller ports to physical ports.
- * 
- * Ex. gVirtualToPhysicalContPorts[virtualPort] = physicalPort
- */
-extern u8 gVirtualContPortMap[MAXCONTROLLERS]; 
-
-/**
- * Internal joystick X values per controller from last input frame used to set gMenuJoyXSign.
- * 
- * @see gMenuJoyXSign, controller_thread_entry
- */ 
-extern s8 gLastJoyX[MAXCONTROLLERS];
-/**
- * Internal joystick Y values per controller from last input frame used to set gMenuJoyYSign.
- * 
- * @see gMenuJoyYSign, controller_thread_entry 
- */ 
-extern s8 gLastJoyY[MAXCONTROLLERS];
-/**
- * Increments from [0, gMenuJoystickDelay) while abs(joystick X) > 35.
- * Once it reaches gMenuJoystickDelay - 1, resets back to 0.
- */
-extern s8 gMenuJoyXHoldTimer[MAXCONTROLLERS];
-/**
- * Increments from [0, gMenuJoystickDelay) while abs(joystick Y) > 35.
- * Once it reaches gMenuJoystickDelay - 1, resets back to 0.
- */
-extern s8 gMenuJoyYHoldTimer[MAXCONTROLLERS];
-/**
- * For each controller, a value indicating that the joystick X axis is pushed
- * in a direction and that for the current frame it should result in X-axis menu movement.
- * 
- * @details The joystick must be held in a direction for gMenuJoystickDelay frames
- * before the sign will be non-zero here, and it will remain non-zero just for the
- * one frame until the stick is held for another gMenuJoystickDelay frames.
- * This effectively allows the user to hold the joystick in a direction to move between
- * menus without it being too fast.
- * 
- * @see gMenuJoystickDelay, gMenuJoyXHoldTimer
- */
-extern s8 gMenuJoyXSign[MAXCONTROLLERS];
-/**
- * For each controller, a value indicating that the joystick Y axis is pushed
- * in a direction and that for the current frame it should result in Y-axis menu movement.
- * 
- * @details The joystick must be held in a direction for gMenuJoystickDelay frames
- * before the sign will be non-zero here, and it will remain non-zero just for the
- * one frame until the stick is held for another gMenuJoystickDelay frames.
- * This effectively allows the user to hold the joystick in a direction to move between
- * menus without it being too fast.
- * 
- * @see gMenuJoystickDelay, gMenuJoyYHoldTimer
- */
-extern s8 gMenuJoyYSign[MAXCONTROLLERS];
-
-extern OSThread gControllerThread;
-extern u8 gControllerThreadStack[CONTROLLER_THREAD_STACKSIZE];
-
-extern OSScClient gContSchedulerClient;
-/**
- * The address of this message is sent to gContThreadMesgQueue.
- * 
- * @see gContThreadMesgQueue for more info.
- */
-extern s16 gContQueue2Message;
-
-/**
- * The number of controller input frames that a joystick axis must be held
- * before menu movement will automatically occur. 
- * 
- * For ex. if the user move the joystick down, the menu selection will move
- * instantly initially but then wait this many input frames before moving again
- * while the joystick is held.
- * 
- * Defaults to 5.
- */ 
-extern s8 gMenuJoystickDelay;
-
-void controller_thread_entry(void *arg);
-s8 handle_joystick_deadzone(s8 stick);
-void init_virtual_cont_port_map();
 
 /**
  * @returns The message queue associated with SI controller interrupts.
@@ -264,10 +14,10 @@ void signal_apply_controller_inputs() {
 }
 
 /**
- * Signals the controller thread to apply controller inputs to input related globals 
- * used by normal gameplay functions the next time it handles controller input and 
+ * Signals the controller thread to apply controller inputs to input related globals
+ * used by normal gameplay functions the next time it handles controller input and
  * blocks until this occurs.
- * 
+ *
  * @see gContThreadMesgQueue
  */
 void apply_controller_inputs() {
@@ -289,14 +39,14 @@ s32 init_controller_data() {
     //
     // osContInit requires a message queue associated with SI interrupts
     osCreateMesgQueue(
-        &gContInterruptQueue, 
+        &gContInterruptQueue,
         &gContInterruptBuffer[0],
         CONT_INTERRUPT_BUFFER_LENGTH);
 
     osSetEventMesg(OS_EVENT_SI, &gContInterruptQueue, gContInterruptMessage);
 
     osContInit(&gContInterruptQueue, &contBitpattern, &gContStatuses[0]);
-    
+
     // Start reading controller input
     osContStartReadData(&gContInterruptQueue);
 
@@ -315,10 +65,10 @@ s32 init_controller_data() {
         gLastJoyX[i] = gLastJoyY[i] = 0;
         gMenuJoyXHoldTimer[i] = gMenuJoyYHoldTimer[i] = 0;
         gMenuJoyXSign[i] = gMenuJoyYSign[i] = 0;
-        
+
         gButtonPresses[i] = 0;
         gButtonReleases[i] = 0;
-        
+
         // Note: Must check bitpattern and errno to determine if a controller is inserted
         if ((contBitpattern & (1 << i)) && !(gContStatuses[i].errno & CONT_NO_RESPONSE_ERROR)) {
             lastControllerIndex = i;
@@ -334,26 +84,26 @@ s32 init_controller_data() {
 
 void start_controller_thread(OSSched *scheduler) {
     osCreateMesgQueue(
-        /*mq*/      &gContThreadMesgQueue, 
-        /*msg*/     &gContThreadMesgQueueBuffer[0], 
+        /*mq*/      &gContThreadMesgQueue,
+        /*msg*/     &gContThreadMesgQueueBuffer[0],
         /*count*/   CONT_THREAD_MESG_QUEUE_BUFFER_LENGTH
     );
 
     osScAddClient(scheduler, &gContSchedulerClient, &gContThreadMesgQueue, 2);
 
     osCreateMesgQueue(
-        /*mq*/      &gContThreadInputsAppliedQueue, 
-        /*msg*/     &gContThreadInputsAppliedQueueBuffer[0], 
+        /*mq*/      &gContThreadInputsAppliedQueue,
+        /*msg*/     &gContThreadInputsAppliedQueueBuffer[0],
         /*count*/   CONT_THREAD_INPUTS_APPLIED_QUEUE_BUFFER_LENGTH
     );
 
     // Create and start controller thread
     osCreateThread(
-        /*t*/       &gControllerThread, 
-        /*id*/      CONTROLLER_THREAD_ID, 
-        /*entry*/   &controller_thread_entry, 
-        /*arg*/     NULL, 
-        /*sp*/      &gControllerThreadStack[CONTROLLER_THREAD_STACKSIZE], 
+        /*t*/       &gControllerThread,
+        /*id*/      CONTROLLER_THREAD_ID,
+        /*entry*/   &controller_thread_entry,
+        /*arg*/     NULL,
+        /*sp*/      &gControllerThreadStack[CONTROLLER_THREAD_STACKSIZE],
         /*pri*/     12
     );
 
@@ -374,7 +124,7 @@ void controller_thread_entry(void *_) {
 
     int totalStickX;
     int totalStickY;
-    
+
     ControllersSnapshot *curSnaps;
     ControllersSnapshot *nextSnaps;
     ControllersSnapshot *nextSnap;
@@ -396,7 +146,7 @@ void controller_thread_entry(void *_) {
 
         if (*contMesg != 1) {
             if (*contMesg != 0xA) { // ???
-                
+
             }
 
             // Next time a 1 is received, apply inputs after collecting interrupts
@@ -571,10 +321,10 @@ void _init_virtual_cont_port_map() {
  * TODO: Not sure what the intent of this function is, but here's some example in/outs:
  * ports = [3, 0, 2, 1]
  * gVirtualContPortMap = [0, 2, 3, 1]
- * 
+ *
  * ports = [0, 1, 2, 3]
  * gVirtualContPortMap = [1, 2, 3, 0]
- * 
+ *
  * ports = [1, 3, 2, 0]
  * gVirtualContPortMap = [0, 1, 2, 3]
  */
@@ -627,7 +377,7 @@ u16 get_masked_buttons(int port) {
 u16 get_masked_buttons_from_buffer(int port, int buffer) {
     OSContPad *pads;
     ControllersSnapshot *snapshot;
-    
+
     if (port > 0) {
         return 0;
     }
@@ -664,7 +414,7 @@ u16 get_button_presses(int port) {
 
 u16 get_masked_button_presses_from_buffer(int port, int buffer) {
     ControllersSnapshot *snapshot;
-    
+
     if (port > 0) {
         return 0;
     }
@@ -691,7 +441,7 @@ u16 get_masked_button_releases(int port) {
 
 u16 get_masked_button_releases_from_buffer(int port, int buffer) {
     ControllersSnapshot *snapshot;
-    
+
     if (port > 0) {
         return 0;
     }
@@ -719,7 +469,7 @@ s8 get_joystick_x(int port) {
 s8 get_joystick_x_from_buffer(int port, int buffer) {
     OSContPad *pads;
     ControllersSnapshot *snapshot;
-    
+
     if (port > 0) {
         return 0;
     }
@@ -749,7 +499,7 @@ s8 get_joystick_y(int port) {
 s8 get_joystick_y_from_buffer(int port, int buffer) {
     OSContPad *pads;
     ControllersSnapshot *snapshot;
-    
+
     if (port > 0) {
         return 0;
     }
@@ -775,14 +525,14 @@ void get_joystick_menu_xy_sign(int port, s8 *xSign, s8 *ySign) {
 
 /**
  * Applies a lower deadzone of -8,8 and upper deadzone of -70,70 to a joystick axis input.
- * 
+ *
  * For example:
  * - If stick is between -8 and 8, the returned value is 0
  * - If stick is greater than 8 or less than -8, the returned value will be moved towards
  *   zero by 8 and clamped at an upper value of 70 and a lower value of -70.
- * 
+ *
  * If gIgnoreJoystick is TRUE, this will always return 0.
- * 
+ *
  * Returns a joystick axis value between [-70, 70].
  */
 s8 handle_joystick_deadzone(s8 stick) {
